% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/FrissMessageBox.R
\name{addMessageBoxToPage}
\alias{addMessageBoxToPage}
\title{Adds includes for messagebox to page Should be called once on top of ui.r}
\usage{
addMessageBoxToPage(style = "message_default.css")
}
\arguments{
\item{style}{Stylesheet to use for the messagebox. Messagebox.js default comes with:
message_default.css
message_growl_dark.css
message_growl_shiny.css
message_skyblue.css
message_solid.css}
}
\description{
Adds includes for messagebox to page Should be called once on top of ui.r
}
\keyword{includes}
\keyword{messagebox.js}

