% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/FrissMessageBox.R
\name{addMessageHandler}
\alias{addMessageHandler}
\title{Adds javascript message handler to the page}
\usage{
addMessageHandler()
}
\description{
Adds javascript message handler to the page
}
\keyword{handler}
\keyword{javascript}
\keyword{message}

