% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/FrissMessageBox.R
\name{hideSplash}
\alias{hideSplash}
\title{Hides a spalsh screen created with showSplash.}
\usage{
hideSplash(session, formName = "splash")
}
\arguments{
\item{session}{Shiny session object from the server.r page of the app that will use this function}

\item{formName}{Id of the form to hide}
}
\description{
Hides a spalsh screen created with showSplash.
}
\keyword{alert}
\keyword{dhtmlx}
\keyword{hide}
\keyword{modal}
\keyword{popup}
\keyword{splash}

