% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/FrissMessageBox.R
\name{showConfirm}
\alias{showConfirm}
\title{Shows a confirm message.}
\usage{
showConfirm(session, inputName, message, okText = "OK",
  cancelText = "Cancel")
}
\arguments{
\item{session}{Shiny session object from the server.r page of the app that will use this function}

\item{inputName}{Name of the input to return the result to.}

\item{message}{Message to display on the confirm box}
}
\description{
Shows a confirm message in the browser and returns result in input$inputName
}
\keyword{alert}
\keyword{confirm}
\keyword{dhtmlx}
\keyword{modal}
\keyword{popup}
\keyword{show}

