% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/FrissMessageBox.R
\name{showFormModal}
\alias{showFormModal}
\title{Shows a modal with custom html in it}
\usage{
showFormModal(session, inputName, html, okText = "ok",
  cancelText = "cancel", modalStyle = "display:none",
  buttonStyle = "width:50x;padding:")
}
\arguments{
\item{session}{Shiny session object from the server.r page of the app that will use this function.}

\item{inputName}{Name of the input to return the result to.}

\item{html}{Html to put on the modal}

\item{okText}{Text to put on the ok button}

\item{cancelText}{Text to put on the cancel button}

\item{modalStyle}{Style sheet properties of the modal}

\item{buttonStyle}{Style sheet properties of the button}
}
\description{
Shows a modal with custom html in it and sends result back to input$inputName when an ok / cancel button is pressed
}
\keyword{alert}
\keyword{confirm}
\keyword{dhtmlx}
\keyword{modal}
\keyword{popup}
\keyword{show}

