% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/FrissMessageBox.R
\name{showMessage}
\alias{showMessage}
\title{Shows a temporary message.}
\usage{
showMessage(session, message, type = "message")
}
\arguments{
\item{session}{Shiny session object from the server.r page of the app that will use this function.}

\item{message}{Message to display.}

\item{type}{On of the perdefined message box types (see http://docs.dhtmlx.com/message__working_with_windows.html) or a custom css class}
}
\description{
Shows a temporary message on the right of the screen.
}
\keyword{alert}
\keyword{confirm}
\keyword{dhtmlx}
\keyword{modal}
\keyword{popup}
\keyword{show}

