% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/FrissMessageBox.R
\name{showPopup}
\alias{showPopup}
\title{shows simple pop up message}
\usage{
showPopup(session, message, title = "", okText = "OK")
}
\arguments{
\item{session}{Shiny session object from the server.r page of the app that will use this function}

\item{message}{Message to display in the box}

\item{title}{Title of the messagebox}

\item{okText}{Text to put on the 'OK' button. Default is 'OK'}
}
\description{
This function shows a simple dhtmlx alert
}
\keyword{alert}
\keyword{dhtmlx}
\keyword{modal}
\keyword{popup}

