% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/FrissMessageBox.R
\name{showSplash}
\alias{showSplash}
\title{shows a simple splash screen}
\usage{
showSplash(session, html = "<h1>Please wait...</h1>", formName = "splash")
}
\arguments{
\item{session}{Shiny session object from the server.r page of the app that will use this function}

\item{html}{Html code to put on the splash screen}

\item{formName}{Id of the splash screen. Used to determine if a screen is already created.}
}
\description{
shows splash screen. This screen will remain visible and on top until hideSplash is called
}
\keyword{alert}
\keyword{dhtmlx}
\keyword{modal}
\keyword{popup}
\keyword{splash}

